feature: modularized list implementation
[#42] https://github.com/kinten108101/steam-vpk/issues/42
