+++ data/com.github.kinten108101.SteamVpk.data.gresource.xml
@@ -2,9 +2,9 @@
 <gresources>
 	<gresource prefix="/com/github/kinten108101/SteamVpk">
     <file>css/style.css</file>
-	  <file>ui/addonlist-list-box.ui</file>
     <file>ui/addonlist-page.ui</file>
 	  <file>ui/addonlist-row.ui</file>
+    <file>ui/addonlist-section.ui</file>
     <file>ui/download-page.ui</file>
     <file>ui/main-window.ui</file>
     <file>ui/omnibar-popover.ui</file>
