--- data/ui/addonlist-list-box.blp
@@ -1,11 +1,0 @@
-using Gtk 4.0;
-using Adw 1;
-
-template AddonlistListBox : Gtk.ListBox {
-
-  selection-mode: none;
-  valign: start;
-
-  styles ["boxed-list"]
-}
-
