+++ src/sample_addons.ts
@@ -1,0 +1,124 @@
+import { IAddonlistItem } from './addonlist-model.js';
+
+export const sample_addons: IAddonlistItem[] = [{
+  name: 'INS:S Louis',
+  id: 'insslouis@mav',
+  description:
+`punk ass island biÿtch
+
+Louis dressed in weekend clothing and a tactical vest, from Insurgency Sandstorm.
+Features remade UI elements, first person hands, and game accurate models and materials.
+
+Credits:
+Nickolox: Painted over portraits and character icons
+YuRaNnNzZZ, TFA, kuma: Materialization Method
+mav: Model Porting, Project Lead
+
+Note:
+You are not allowed to modify, use parts of and create deriative works from contents of this addon without permission. Reuploading any parts of the addon is a direct violation of Steam Workshop Rules, Steam Subscriber Agreement and Steam Online Conduct.`,
+  last_update: '11 Nov 2019',
+},
+{
+  name: 'Hospital Survivor - Bill',
+  id: 'bill-hospital@Someone',
+  description:
+`Another crappy mod idea of mine. Try to enjoy
+
+This mod includes:
+1.Model itself
+2.Boomer texture
+3.Default vgui
+4.'Small' adjustment to fps hands
+5.Facials
+6.Should work with amy Bill retexture
+7. Shoots and fights with just one hand (insert shock emoji here)
+8.Passing model
+
+The models i used:
+Model 1 click here[sketchfab.com]
+Model 2 click here[sketchfab.com]
+Model 3 click here[sketchfab.com]
+
+Also open for commissions, yeah`,
+  last_update: '30 Mar @ 11:18pm',
+},
+{
+  name: 'EFT Ellis bear zasion',
+  id: 'ellis-tactical@Animal33',
+  description:
+`替换Ellis。
+
+Model and material：Battlestate Games
+Binding mold：Animal33
+material handling：BaiF*ckmouse
+HUD：Animal33`,
+  last_update: '17 Mar @ 1:08pm',
+},
+{
+  name: 'MW2019 FR 5.56',
+  id: 'mw2019-fr556@mav',
+  description:
+`3 round burst bullpup battle rifle. A well placed burst can be extremely deadly at intermittent range.
+
+The FR 5.56 cleaned up. Features fully fixed and accurate game animations, sounds, and materials.
+
+Credits:
+Scobalula: Greyhound
+YuRaNnNzZZ, TFA, FlamingFox: Materialization Method and Assistance
+ThomasCat: Conversion Rig, Animation Porting
+mav: Animation Porting, Model Porting, Sound Porting, Project Lead
+Biddin: Incendiary Sound
+
+Note:
+You are not allowed to modify, use parts of and create deriative works from contents of this addon without permission. Reuploading any parts of the addon is a direct violation of Steam Workshop Rules, Steam Subscriber Agreement and Steam Online Conduct.`,
+  last_update: '6 May, 2022 @ 3:12am',
+},
+{
+  name: 'COD:MW FAMAS F1 SD Black',
+  id: 'mw19-famas-sd@denny',
+  description:
+`Happy New Year!!!!:D
+
+Replaces Desert Rifle/SCAR
+替换SCAR三连发步枪
+
+Yea yea my compile.
+
+Credits:
+
+Infinity Ward - Famas models/textures/Audio
+Denny凯妈/Infinity Ward/zmg - Audio
+Viper - Original Compile/Ripping/textures/Animations
+70R3H - Textures
+mav - runidle Animations
+Denny凯妈/zmg - Compile/Audio/Texture editing/Animation editing/Fixing
+It includes a HUD icon:)
+Happy new year! Happy every day!
+Enjoy it!
+
+Original part to Garry'mod by Viper click ↓`,
+  last_update: '1 Jan @ 8:54am',
+},
+{
+  name: 'MW2019 Ellis',
+  id: 'mw2019-ellis@Animal33',
+  description:
+`模型：Modern Warfare2019
+
+
+
+替换-Ellis
+
+感谢MLUI的表情修复
+
+
+
+
+
+---------------------------------------------------------------------
+
+喜欢的也是可以点个赞和收藏`,
+  last_update: '8 Sep, 2022 @ 7:30pm',
+  in_randomizer: false,
+},
+];
